make unit tests pass
